Move env inheritence into osutils and have installer use it.
This allows the installer shell to inherit the existing environment.
